batch-containers/requirements.txt
azure-batch==5.1.0
